cleaned up the files to resolve merge conflicts
