+++ front-end/app/_layout.tsx
@@ -1,12 +1,12 @@
-// import { Stack } from 'expo-router';
+import { Stack } from 'expo-router';
 
-// export default function Layout() {
-//   return (
-//     <Stack>
-//       <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-//     </Stack>
-//   );
-// }
+export default function Layout() {
+  return (
+    <Stack>
+      <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+    </Stack>
+  );
+}
 
 
 // import {Tabs} from "expo-router";
